refactor(quotes): simplify SimpleQuote and share the Flag test helper
